defaults to v0.10.0.0. #63
defaulting to v0.10.1.0 breaks client.GetOffset against brokers that are older
than v0.10.1.0. i thought of fixing this by detecting the broker version
automatically #63. but that is a bit overkill at the moment imo, as support for
ApiVersionsRequest was only added in v0.10.0.0 - so the only distinction the
automatic resolver could make is between v0.10.0.0 and v0.10.1.0 - all other
versioning happened before v0.10.0.0. the additional features around time based
offset searching are not yet supported in kt, so there's no need to require
v0.10.1.0 at this point.