understand time-based positions in parseOffsets
This is a preparatory step before implementing full time-based
queries in kt. We change the offsets syntax so that it allows
time-based positions, including duration-based relative
positions. To avoid the grammatical conflict from the presence
of colons in time representations, times are enclosed in square
brackets.

Adding times makes the current "offset" type name a little
awkward, as it's hard to distinguish between "offset" as a
way of describing a numerical offset and "offset" as a term
for a description of a position within a stream. To make that clearer
in the code, we rename the `offset` type to `position`.
We also refactor that type so that it can represent both
time and offset-based start and differences, and to remove the
redundant isRelative boolean flag (OffsetOldest and friends are
more "symbolic" than "relative" anyway).

To enable straightforward comparison of position values without
using DeepEquals (which is problematic with time.Time), we
change `TestParseOffsets` to use the quicktest package. There
may be some advantage to using that package for testing throughout
in the future for self-consistency and better test error reports.