use testscript for system tests
This makes the system test cleaner (it doesn't need to build another binary, for example), and arguably more readable.
It also paves the way for more end-to-end tests.

Exactly the same sequence of commands is being tested as before, except that some extra output checking is done where there wasn't any before.