print consumed messages in time order
Currently when consuming from multiple partitions, messages are
arbitrarily interleaved. Although there is no inherent ordering
between partitions, it's nice to see them in (approximate) time order,
so this PR prints messages in timestamp order as far as possible.

In `-f` (follow) mode, messages are still printed as soon as they arrive.

As a nice side-effect, this means that output across partitions
can be deterministic, opening up the way to having system-level
tests that test output across multiple partitions.

We also add the capability for `hkt produce` to set a timestamp on
produced messages.

We also change the name of the timestamp field printed by
the consumer to `time` from `timestamp` for consistency with
the `hkt produce` command.