simpler printing
We don't need a separate goroutine to print to stdout - a simple
printer is sufficient, and then we don't need to create
ephemeral channels to know when the printing is done
or manage the lifetime of the printer goroutine.

We might not even need the mutex, but we'll keep it around
for now, as printing concurrently to stdout isn't necessarily
safe on all platforms.