JSON values for messages
We add a "json" type for consumed message values, causing messages
to be priunted as the actual JSON rather than quoting it.
The `produce` command also gets support for this.

This is now the default encoding type for message values.

I found myself unable to remember the flag name (quick! does `consume`
take `-encodevalue` or `-decodevalue`?), so I've changed the
flag to `-valuecodec` and `-keycodec`, common between
`hkt consume` and `hkt produce`.