implement key filtering
It's common to search for messages with a particular key, and
since the relationship between key and partition tends to be stable,
this can be made more efficient by finding what partition the key
corresponds to and reading messages from just that partition.

This PR implements that by adding a `-key` flag to the `hkt consume`
command. See the consume help text for details.