use standard Sarama producer to produce messages
This simplifies the code and makes the message partitioning
consistent with the standard Sarama partitioning,
which will enable us to test the up-coming key selection
feature more easily.

We remove the `-partition` flag to the produce command,
because it's possible to specify it with the `partition` field
of the input messages themselves, and usually for messages
without a key, we want the partitioner to choose the partition.
In the future, we could allow a number for the partitioner, for
example `-partitioner=6` to specify a default partition.

We also change the producer so that it doesn't print a line
to the standard output for every message produced. Silence is
golden.